refactor: 타이머 로직을 useCountdown 훅으로 분리
