+++ frontend/src/features/discussion/ui/DiscussionInviteCard/TimerButton.tsx
@@ -1,33 +1,15 @@
-import { useEffect, useState } from 'react';
-
 import Button from '@/components/Button';
-import { SECOND_IN_MILLISECONDS } from '@/utils/date';
+import useCountdown from '@/hooks/useCountdown';
 
 import { timerButtonStyle } from './index.css';
 
 interface TimerButtonProps {
   targetDateTime: Date;
-  onTimeEnd?: () => void;
+  onTimeEnd: () => void;
 }
 
-const TimerButton = ({ targetDateTime: targetTime, onTimeEnd }: TimerButtonProps) => {
-  const [remainingTime, setRemainingTime] = useState<number>(
-    Math.max(targetTime.getTime() - Date.now(), 0),
-  );
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      const remaining = Math.max(targetTime.getTime() - Date.now(), 0);
-      setRemainingTime(remaining);
-      if (remaining === 0) {
-        clearInterval(interval);
-        if (onTimeEnd) onTimeEnd();
-      }
-    }, SECOND_IN_MILLISECONDS);
-
-    return () => clearInterval(interval);
-  }, [targetTime, onTimeEnd]);
-
+const TimerButton = ({ targetDateTime, onTimeEnd }: TimerButtonProps) => {
+  const remainingTime = useCountdown(targetDateTime, onTimeEnd);
   return (
     <Button
       className={timerButtonStyle}
